Emscripten:  Implemented iupdrvGetPreferencePath to get things building again. IupConfig files are NOT persistent between runs due to the way Emscripten/web works. This will require more changes to handle.
TODO: Emscripten virtualizes everything which means this data is not persistent.
A browser reload will lose all data.
There is a thing called IDBFS which writes to a persistent IndexedDB,
however, the integration is not obvious and will require research.
I suspect we will need to look at the configuratin file's actual initial load and final write to disk and rewrite those sections.