Emscripten: Implemented resize and basic layout (including Create condition). Because of the apparent complexity of innerWidth vs. clientWidth, I'm using a helper function from https://github.com/tysonmatanich/viewportSize/blob/master/viewportSize.js, Tyson Matanich, 2013 | License: MIT. Also, to make the internal layout work, I set the ih->currentwidth/height to use inner/client instead of outer. We may need to scrutinize where APIs expect the outer instead.
I also added a DIALOG check in the Common base layout routine. In experimenting trying to get things working (I had a bug), I ended up causing the dialog to go through the base layout and it throws and exception because there is no elem.style in the dialog. I have since reverted those changes that caused the dialog to go through that code, but as a future safety check, I decided to leave the checks there.

I also renamed the JS internal helper namespace from Dialog to IupDialog.